[cherry-pick] merge `main` into `sui-v29.0-release` (#9892)
## Test Plan 

Verified none of them require regenesis

---------
