+++ crates/sui-adapter-transactional-tests/tests/sui/coin_in_vec.exp
@@ -1,0 +1,15 @@
+processed 4 tasks
+
+init:
+A: object(103)
+
+task 1 'publish'. lines 6-32:
+created: object(105), object(106)
+written: object(104)
+
+task 2 'run'. lines 34-34:
+written: object(105), object(107)
+deleted: object(104)
+
+task 3 'run'. lines 36-36:
+written: object(104), object(105), object(108)
