+++ crates/sui-indexer/src/apis/mod.rs
@@ -2,17 +2,19 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub(crate) use coin_api::CoinReadApi;
-pub(crate) use event_api::EventReadApi;
 pub(crate) use extended_api::ExtendedApi;
 pub(crate) use governance_api::GovernanceReadApi;
+pub(crate) use indexer_api::IndexerApi;
+pub(crate) use move_utils::MoveUtilsApi;
 pub(crate) use read_api::ReadApi;
 pub(crate) use transaction_builder_api::TransactionBuilderApi;
 pub(crate) use write_api::WriteApi;
 
 mod coin_api;
-mod event_api;
 mod extended_api;
 mod governance_api;
+mod indexer_api;
+mod move_utils;
 mod read_api;
 mod transaction_builder_api;
 mod write_api;
