+++ crates/sui-indexer/src/apis/read_api.rs
@@ -1,30 +1,26 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::errors::IndexerError;
-use crate::store::IndexerStore;
-use crate::types::{SuiTransactionFullResponse, SuiTransactionFullResponseWithOptions};
 use async_trait::async_trait;
 use futures::future::join_all;
 use jsonrpsee::core::RpcResult;
 use jsonrpsee::http_client::HttpClient;
 use jsonrpsee::RpcModule;
-use std::collections::BTreeMap;
-use sui_json_rpc::api::{validate_limit, ReadApiClient, ReadApiServer, QUERY_MAX_RESULT_LIMIT};
+
+use sui_json_rpc::api::{ReadApiClient, ReadApiServer};
 use sui_json_rpc::SuiRpcModule;
 use sui_json_rpc_types::{
-    BigInt, Checkpoint, CheckpointId, CheckpointPage, DynamicFieldPage, MoveFunctionArgType,
-    ObjectsPage, Page, SuiCheckpointSequenceNumber, SuiGetPastObjectRequest,
-    SuiMoveNormalizedFunction, SuiMoveNormalizedModule, SuiMoveNormalizedStruct,
-    SuiObjectDataOptions, SuiObjectResponse, SuiObjectResponseQuery, SuiPastObjectResponse,
-    SuiTransactionResponse, SuiTransactionResponseOptions, SuiTransactionResponseQuery,
-    TransactionsPage,
+    BigInt, Checkpoint, CheckpointId, CheckpointPage, SuiCheckpointSequenceNumber, SuiEvent,
+    SuiGetPastObjectRequest, SuiObjectDataOptions, SuiObjectResponse, SuiPastObjectResponse,
+    SuiTransactionResponse, SuiTransactionResponseOptions,
 };
 use sui_open_rpc::Module;
-use sui_types::base_types::{ObjectID, SequenceNumber, SuiAddress, TxSequenceNumber};
+use sui_types::base_types::{ObjectID, SequenceNumber, TxSequenceNumber};
 use sui_types::digests::TransactionDigest;
-use sui_types::dynamic_field::DynamicFieldName;
-use sui_types::query::TransactionFilter;
+
+use crate::errors::IndexerError;
+use crate::store::IndexerStore;
+use crate::types::{SuiTransactionFullResponse, SuiTransactionFullResponseWithOptions};
 
 pub(crate) struct ReadApi<S> {
     fullnode: HttpClient,
@@ -106,155 +102,6 @@ impl<S: IndexerStore> ReadApi<S> {
         Ok(tx_resp_vec)
     }
 
-    fn query_transactions_internal(
-        &self,
-        query: SuiTransactionResponseQuery,
-        cursor: Option<TransactionDigest>,
-        limit: Option<usize>,
-        descending_order: Option<bool>,
-    ) -> Result<TransactionsPage, IndexerError> {
-        let limit = validate_limit(limit, QUERY_MAX_RESULT_LIMIT)?;
-        let is_descending = descending_order.unwrap_or_default();
-        let cursor_str = cursor.map(|digest| digest.to_string());
-
-        let opts = query.options.unwrap_or_default();
-        if !opts.only_digest() {
-            // TODO(chris): implement this as a separate PR
-            return Err(IndexerError::NotImplementedError(
-                "options has not been implemented on indexer for queryTransactions".to_string(),
-            ));
-        }
-
-        let digests_from_db = match query.filter {
-            None => {
-                let indexer_seq_number = self
-                    .state
-                    .get_transaction_sequence_by_digest(cursor_str, is_descending)?;
-                self.state.get_all_transaction_digest_page(
-                    indexer_seq_number,
-                    limit + 1,
-                    is_descending,
-                )
-            }
-            Some(TransactionFilter::MoveFunction {
-                package,
-                module,
-                function,
-            }) => {
-                let move_call_seq_number = self
-                    .state
-                    .get_move_call_sequence_by_digest(cursor_str, is_descending)?;
-                self.state.get_transaction_digest_page_by_move_call(
-                    package.to_string(),
-                    module,
-                    function,
-                    move_call_seq_number,
-                    limit + 1,
-                    is_descending,
-                )
-            }
-            Some(TransactionFilter::InputObject(input_obj_id)) => {
-                let input_obj_seq = self
-                    .state
-                    .get_input_object_sequence_by_digest(cursor_str, is_descending)?;
-                self.state.get_transaction_digest_page_by_input_object(
-                    input_obj_id.to_string(),
-                    /* version */ None,
-                    input_obj_seq,
-                    limit + 1,
-                    is_descending,
-                )
-            }
-            Some(TransactionFilter::ChangedObject(mutated_obj_id)) => {
-                let indexer_seq_number = self
-                    .state
-                    .get_transaction_sequence_by_digest(cursor_str, is_descending)?;
-                self.state.get_transaction_digest_page_by_mutated_object(
-                    mutated_obj_id.to_string(),
-                    indexer_seq_number,
-                    limit + 1,
-                    is_descending,
-                )
-            }
-            // NOTE: more efficient to run this query over transactions table
-            Some(TransactionFilter::FromAddress(sender_address)) => {
-                let indexer_seq_number = self
-                    .state
-                    .get_transaction_sequence_by_digest(cursor_str, is_descending)?;
-                self.state.get_transaction_digest_page_by_sender_address(
-                    sender_address.to_string(),
-                    indexer_seq_number,
-                    limit + 1,
-                    is_descending,
-                )
-            }
-            Some(TransactionFilter::ToAddress(recipient_address)) => {
-                let recipient_seq_number = self
-                    .state
-                    .get_recipient_sequence_by_digest(cursor_str, is_descending)?;
-                self.state
-                    .get_transaction_digest_page_by_sender_recipient_address(
-                        /* from */ None,
-                        recipient_address.to_string(),
-                        recipient_seq_number,
-                        limit + 1,
-                        is_descending,
-                    )
-            }
-            Some(TransactionFilter::FromAndToAddress { from, to }) => {
-                let recipient_seq_number = self
-                    .state
-                    .get_recipient_sequence_by_digest(cursor_str, is_descending)?;
-                self.state
-                    .get_transaction_digest_page_by_sender_recipient_address(
-                        Some(from.to_string()),
-                        to.to_string(),
-                        recipient_seq_number,
-                        limit + 1,
-                        is_descending,
-                    )
-            }
-            Some(TransactionFilter::TransactionKind(tx_kind_name)) => {
-                let indexer_seq_number = self
-                    .state
-                    .get_transaction_sequence_by_digest(cursor_str, is_descending)?;
-                self.state.get_transaction_digest_page_by_transaction_kind(
-                    tx_kind_name,
-                    indexer_seq_number,
-                    limit + 1,
-                    is_descending,
-                )
-            }
-        }?;
-
-        // digests here are of size (limit + 1), where the last one is the cursor for the next page
-        let mut tx_digests = digests_from_db
-            .iter()
-            .map(|digest| {
-                let tx_digest: Result<TransactionDigest, _> = digest.clone().parse();
-                tx_digest.map_err(|e| {
-                    IndexerError::SerdeError(format!(
-                        "Failed to deserialize transaction digest: {:?} with error {:?}",
-                        digest, e
-                    ))
-                })
-            })
-            .collect::<Result<Vec<TransactionDigest>, IndexerError>>()?;
-
-        let has_next_page = tx_digests.len() > limit;
-        tx_digests.truncate(limit);
-        let next_cursor = tx_digests.last().cloned().map_or(cursor, Some);
-
-        Ok(Page {
-            data: tx_digests
-                .into_iter()
-                .map(SuiTransactionResponse::new)
-                .collect(),
-            next_cursor,
-            has_next_page,
-        })
-    }
-
     fn get_object_with_options_internal(
         &self,
         object_id: ObjectID,
@@ -276,20 +123,7 @@ impl<S> ReadApiServer for ReadApi<S>
 where
     S: IndexerStore + Sync + Send + 'static,
 {
-    async fn get_owned_objects(
-        &self,
-        address: SuiAddress,
-        query: Option<SuiObjectResponseQuery>,
-        cursor: Option<ObjectID>,
-        limit: Option<usize>,
-        at_checkpoint: Option<CheckpointId>,
-    ) -> RpcResult<ObjectsPage> {
-        self.fullnode
-            .get_owned_objects(address, query, cursor, limit, at_checkpoint)
-            .await
-    }
-
-    async fn get_object_with_options(
+    async fn get_object(
         &self,
         object_id: ObjectID,
         options: Option<SuiObjectDataOptions>,
@@ -298,45 +132,18 @@ where
             .migrated_methods
             .contains(&"get_object_with_options".into())
         {
-            return self
-                .fullnode
-                .get_object_with_options(object_id, options)
-                .await;
+            return self.fullnode.get_object(object_id, options).await;
         }
 
         Ok(self.get_object_with_options_internal(object_id, options)?)
     }
 
-    async fn multi_get_object_with_options(
+    async fn multi_get_objects(
         &self,
         object_ids: Vec<ObjectID>,
         options: Option<SuiObjectDataOptions>,
     ) -> RpcResult<Vec<SuiObjectResponse>> {
-        return self
-            .fullnode
-            .multi_get_object_with_options(object_ids, options)
-            .await;
-    }
-
-    async fn get_dynamic_fields(
-        &self,
-        parent_object_id: ObjectID,
-        cursor: Option<ObjectID>,
-        limit: Option<usize>,
-    ) -> RpcResult<DynamicFieldPage> {
-        self.fullnode
-            .get_dynamic_fields(parent_object_id, cursor, limit)
-            .await
-    }
-
-    async fn get_dynamic_field_object(
-        &self,
-        parent_object_id: ObjectID,
-        name: DynamicFieldName,
-    ) -> RpcResult<SuiObjectResponse> {
-        self.fullnode
-            .get_dynamic_field_object(parent_object_id, name)
-            .await
+        return self.fullnode.multi_get_objects(object_ids, options).await;
     }
 
     async fn get_total_transaction_number(&self) -> RpcResult<BigInt> {
@@ -349,25 +156,6 @@ where
         Ok(self.get_total_transaction_number_internal()?.into())
     }
 
-    async fn query_transactions(
-        &self,
-        query: SuiTransactionResponseQuery,
-        cursor: Option<TransactionDigest>,
-        limit: Option<usize>,
-        descending_order: Option<bool>,
-    ) -> RpcResult<TransactionsPage> {
-        if !self
-            .migrated_methods
-            .contains(&"query_transactions".to_string())
-        {
-            return self
-                .fullnode
-                .query_transactions(query, cursor, limit, descending_order)
-                .await;
-        }
-        Ok(self.query_transactions_internal(query, cursor, limit, descending_order)?)
-    }
-
     async fn get_transactions_in_range_deprecated(
         &self,
         start: TxSequenceNumber,
@@ -378,7 +166,7 @@ where
             .await
     }
 
-    async fn get_transaction_with_options(
+    async fn get_transaction(
         &self,
         digest: TransactionDigest,
         options: Option<SuiTransactionResponseOptions>,
@@ -387,17 +175,14 @@ where
             .migrated_methods
             .contains(&"get_transaction".to_string())
         {
-            return self
-                .fullnode
-                .get_transaction_with_options(digest, options)
-                .await;
+            return self.fullnode.get_transaction(digest, options).await;
         }
         Ok(self
             .get_transaction_with_options_internal(&digest, options)
             .await?)
     }
 
-    async fn multi_get_transactions_with_options(
+    async fn multi_get_transactions(
         &self,
         digests: Vec<TransactionDigest>,
         options: Option<SuiTransactionResponseOptions>,
@@ -406,66 +191,11 @@ where
             .migrated_methods
             .contains(&"multi_get_transactions_with_options".to_string())
         {
-            return self
-                .fullnode
-                .multi_get_transactions_with_options(digests, options)
-                .await;
+            return self.fullnode.multi_get_transactions(digests, options).await;
         }
         Ok(self
             .multi_get_transactions_with_options_internal(&digests, options)
             .await?)
-    }
-
-    async fn get_normalized_move_modules_by_package(
-        &self,
-        package: ObjectID,
-    ) -> RpcResult<BTreeMap<String, SuiMoveNormalizedModule>> {
-        self.fullnode
-            .get_normalized_move_modules_by_package(package)
-            .await
-    }
-
-    async fn get_normalized_move_module(
-        &self,
-        package: ObjectID,
-        module_name: String,
-    ) -> RpcResult<SuiMoveNormalizedModule> {
-        self.fullnode
-            .get_normalized_move_module(package, module_name)
-            .await
-    }
-
-    async fn get_normalized_move_struct(
-        &self,
-        package: ObjectID,
-        module_name: String,
-        struct_name: String,
-    ) -> RpcResult<SuiMoveNormalizedStruct> {
-        self.fullnode
-            .get_normalized_move_struct(package, module_name, struct_name)
-            .await
-    }
-
-    async fn get_normalized_move_function(
-        &self,
-        package: ObjectID,
-        module_name: String,
-        function_name: String,
-    ) -> RpcResult<SuiMoveNormalizedFunction> {
-        self.fullnode
-            .get_normalized_move_function(package, module_name, function_name)
-            .await
-    }
-
-    async fn get_move_function_arg_types(
-        &self,
-        package: ObjectID,
-        module: String,
-        function: String,
-    ) -> RpcResult<Vec<MoveFunctionArgType>> {
-        self.fullnode
-            .get_move_function_arg_types(package, module, function)
-            .await
     }
 
     async fn try_get_past_object(
@@ -523,6 +253,10 @@ where
             .fullnode
             .get_checkpoints(cursor, limit, descending_order)
             .await;
+    }
+
+    async fn get_events(&self, transaction_digest: TransactionDigest) -> RpcResult<Vec<SuiEvent>> {
+        self.fullnode.get_events(transaction_digest).await
     }
 }
