+++ crates/sui-indexer/tests/integration_tests.rs
@@ -25,8 +25,8 @@ pub mod pg_integration_test {
     use sui_indexer::store::{IndexerStore, PgIndexerStore};
     use sui_indexer::test_utils::{start_test_indexer, SuiTransactionResponseBuilder};
     use sui_indexer::{get_pg_pool_connection, new_pg_connection_pool, IndexerConfig};
-    use sui_json_rpc::api::EventReadApiClient;
     use sui_json_rpc::api::ExtendedApiClient;
+    use sui_json_rpc::api::IndexerApiClient;
     use sui_json_rpc::api::{ReadApiClient, TransactionBuilderClient, WriteApiClient};
     use sui_json_rpc_types::{
         BigInt, CheckpointId, EventFilter, SuiMoveObject, SuiObjectData, SuiObjectDataFilter,
@@ -43,9 +43,7 @@ pub mod pg_integration_test {
     use sui_types::query::TransactionFilter;
     use sui_types::utils::to_sender_signed_transaction;
     use test_utils::network::{TestCluster, TestClusterBuilder};
-    use test_utils::transaction::{
-        create_devnet_nft, delete_devnet_nft, publish_nfts_package, transfer_coin,
-    };
+    use test_utils::transaction::{create_devnet_nft, delete_devnet_nft, publish_nfts_package};
 
     const WAIT_UNTIL_TIME_LIMIT: u64 = 60;
 
@@ -122,6 +120,7 @@ pub mod pg_integration_test {
         Ok(tx_response)
     }
 
+    // TODO: we should use SuiClient for tests like this
     async fn execute_simple_transfer(
         test_cluster: &mut TestCluster,
         indexer_rpc_client: &HttpClient,
@@ -139,8 +138,8 @@ pub mod pg_integration_test {
         let gas_objects: Vec<ObjectID> = indexer_rpc_client
             .get_owned_objects(
                 *sender,
-                Some(SuiObjectResponseQuery::new_with_options(
-                    SuiObjectDataOptions::new().with_type(),
+                Some(SuiObjectResponseQuery::new_with_filter(
+                    SuiObjectDataFilter::gas_coin(),
                 )),
                 None,
                 None,
@@ -185,11 +184,11 @@ pub mod pg_integration_test {
             assert!(transaction.is_ok());
             let _fullnode_rpc_tx = test_cluster
                 .rpc_client()
-                .get_transaction_with_options(tx_digest, Some(SuiTransactionResponseOptions::new()))
+                .get_transaction(tx_digest, Some(SuiTransactionResponseOptions::new()))
                 .await
                 .unwrap();
             let _indexer_rpc_tx = indexer_rpc_client
-                .get_transaction_with_options(tx_digest, Some(SuiTransactionResponseOptions::new()))
+                .get_transaction(tx_digest, Some(SuiTransactionResponseOptions::new()))
                 .await
                 .unwrap();
 
@@ -253,90 +252,123 @@ pub mod pg_integration_test {
         Ok(())
     }
 
-    // TODO(chris): re-enable this test after not using devnet_nft
-    // #[tokio::test]
-    // #[timeout(60000)]
-    // async fn test_simple_transaction_e2e() -> Result<(), anyhow::Error> {
-    //     let (mut test_cluster, indexer_rpc_client, store, _handle) = start_test_cluster(None).await;
-    //     // Allow indexer to sync genesis
-    //     wait_until_next_checkpoint(&store).await;
-    //     let (tx_response, sender, recipient, gas_objects) =
-    //         execute_simple_transfer(&mut test_cluster, &indexer_rpc_client).await?;
-    //     wait_until_transaction_synced(&store, tx_response.digest.base58_encode().as_str()).await;
+    #[tokio::test]
+    #[timeout(60000)]
+    async fn test_simple_transaction_e2e() -> Result<(), anyhow::Error> {
+        let (mut test_cluster, indexer_rpc_client, store, _handle) = start_test_cluster(None).await;
+        // Allow indexer to sync genesis
+        wait_until_next_checkpoint(&store).await;
+        let (package_id, publish_digest) =
+            publish_nfts_package(&mut test_cluster.wallet, /* sender */ None).await;
+        wait_until_transaction_synced(&store, publish_digest.base58_encode().as_str()).await;
 
-    //     let (tx_response_2, _, _, _) =
-    //         execute_simple_transfer(&mut test_cluster, &indexer_rpc_client).await?;
-    //     wait_until_transaction_synced(&store, tx_response_2.digest.base58_encode().as_str()).await;
+        let (tx_response, sender, recipient, gas_objects) =
+            execute_simple_transfer(&mut test_cluster, &indexer_rpc_client).await?;
 
-    //     let tx_read_response = indexer_rpc_client
-    //         .get_transaction_with_options(
-    //             tx_response.digest,
-    //             Some(SuiTransactionResponseOptions::full_content()),
-    //         )
-    //         .await?;
-    //     assert_eq!(tx_response.digest, tx_read_response.digest);
-    //     assert_eq!(tx_response.transaction, tx_read_response.transaction);
-    //     assert_eq!(tx_response.effects, tx_read_response.effects);
-    //     assert_eq!(tx_response.events, tx_read_response.events);
-    //     assert_eq!(tx_response.object_changes, tx_read_response.object_changes);
-    //     assert_eq!(
-    //         tx_response.balance_changes,
-    //         tx_read_response.balance_changes
-    //     );
+        wait_until_transaction_synced(&store, tx_response.digest.base58_encode().as_str()).await;
+        let (_, _, nft_digest) = create_devnet_nft(&mut test_cluster.wallet, package_id)
+            .await
+            .unwrap();
+        wait_until_transaction_synced(&store, nft_digest.base58_encode().as_str()).await;
 
-    //     // query tx with sender address
-    //     let from_query =
-    //         SuiTransactionResponseQuery::new_with_filter(TransactionFilter::FromAddress(sender));
-    //     let tx_from_query_response = indexer_rpc_client
-    //         .query_transactions(from_query, None, None, None)
-    //         .await?;
-    //     assert!(!tx_from_query_response.has_next_page);
-    //     assert_eq!(tx_from_query_response.data.len(), 2);
+        let tx_read_response = indexer_rpc_client
+            .get_transaction(
+                tx_response.digest,
+                Some(SuiTransactionResponseOptions::full_content()),
+            )
+            .await?;
+        assert_eq!(tx_response.digest, tx_read_response.digest);
+        assert_eq!(tx_response.transaction, tx_read_response.transaction);
+        assert_eq!(tx_response.effects, tx_read_response.effects);
+        assert_eq!(tx_response.events, tx_read_response.events);
+        assert_eq!(tx_response.object_changes, tx_read_response.object_changes);
+        assert_eq!(
+            tx_response.balance_changes,
+            tx_read_response.balance_changes
+        );
 
-    //     // query tx with recipient address
-    //     let to_query =
-    //         SuiTransactionResponseQuery::new_with_filter(TransactionFilter::ToAddress(recipient));
-    //     let tx_to_query_response = indexer_rpc_client
-    //         .query_transactions(to_query, None, None, None)
-    //         .await?;
-    //     // the address has received 3 transactions, one is genesis
-    //     assert!(!tx_to_query_response.has_next_page);
-    //     assert_eq!(tx_to_query_response.data.len(), 3);
+        // query tx with sender address
+        let from_query =
+            SuiTransactionResponseQuery::new_with_filter(TransactionFilter::FromAddress(sender));
+        let tx_from_query_response = indexer_rpc_client
+            .query_transactions(from_query, None, None, None)
+            .await?;
+        assert!(!tx_from_query_response.has_next_page);
+        assert_eq!(tx_from_query_response.data.len(), 3);
 
-    //     // query tx with mutated object id
-    //     let mutation_query = SuiTransactionResponseQuery::new_with_filter(
-    //         TransactionFilter::ChangedObject(*gas_objects.first().unwrap()),
-    //     );
-    //     let tx_mutation_query_response = indexer_rpc_client
-    //         .query_transactions(mutation_query, None, None, None)
-    //         .await?;
-    //     // the coin is first created by genesis tx, then transferred by the above tx
-    //     assert!(!tx_mutation_query_response.has_next_page);
-    //     assert_eq!(tx_mutation_query_response.data.len(), 2);
-    //     assert_eq!(
-    //         tx_response.digest,
-    //         tx_mutation_query_response.data.last().unwrap().digest,
-    //     );
+        let tx_kind_query = SuiTransactionResponseQuery::new_with_filter(
+            TransactionFilter::TransactionKind("ProgrammableTransaction".to_string()),
+        );
+        let tx_kind_query_response = indexer_rpc_client
+            .query_transactions(tx_kind_query, None, None, None)
+            .await?;
+        assert!(!tx_kind_query_response.has_next_page);
+        assert_eq!(tx_kind_query_response.data.len(), 3);
 
-    //     // query tx with input object id
-    //     let input_query = SuiTransactionResponseQuery::new_with_filter(
-    //         TransactionFilter::InputObject(*gas_objects.first().unwrap()),
-    //     );
-    //     let tx_input_query_response = indexer_rpc_client
-    //         .query_transactions(input_query, None, None, None)
-    //         .await?;
-    //     assert_eq!(tx_input_query_response.data.len(), 1);
-    //     assert_eq!(
-    //         tx_input_query_response.data.first().unwrap().digest,
-    //         tx_response.digest
-    //     );
-    //     assert_eq!(
-    //         Some(tx_input_query_response.data.last().unwrap().digest),
-    //         tx_input_query_response.next_cursor,
-    //     );
-    //     // TODO: implement move call query
-    //     Ok(())
-    // }
+        // query tx with recipient address
+        let to_query =
+            SuiTransactionResponseQuery::new_with_filter(TransactionFilter::ToAddress(recipient));
+        let tx_to_query_response = indexer_rpc_client
+            .query_transactions(to_query, None, None, None)
+            .await?;
+        // the address has received 2 transactions, one is genesis
+        assert!(!tx_to_query_response.has_next_page);
+        assert_eq!(tx_to_query_response.data.len(), 2);
+
+        // query tx with both sender and recipient addresses
+        let from_to_query =
+            SuiTransactionResponseQuery::new_with_filter(TransactionFilter::FromAndToAddress {
+                from: sender,
+                to: recipient,
+            });
+        let tx_from_to_query_response = indexer_rpc_client
+            .query_transactions(from_to_query, None, None, None)
+            .await?;
+        assert!(!tx_from_to_query_response.has_next_page);
+        assert_eq!(tx_from_to_query_response.data.len(), 1);
+        assert_eq!(
+            tx_response.digest,
+            tx_from_to_query_response.data.first().unwrap().digest
+        );
+
+        // query tx with mutated object id
+        let mutation_query = SuiTransactionResponseQuery::new_with_filter(
+            TransactionFilter::ChangedObject(*gas_objects.first().unwrap()),
+        );
+        let tx_mutation_query_response = indexer_rpc_client
+            .query_transactions(mutation_query, None, None, None)
+            .await?;
+        // the coin is first created by genesis tx, then transferred by the above tx
+        assert!(!tx_mutation_query_response.has_next_page);
+        assert_eq!(tx_mutation_query_response.data.len(), 3);
+
+        // query tx with input object id
+        let input_query = SuiTransactionResponseQuery::new_with_filter(
+            TransactionFilter::InputObject(*gas_objects.first().unwrap()),
+        );
+        let tx_input_query_response = indexer_rpc_client
+            .query_transactions(input_query, None, None, None)
+            .await?;
+        assert_eq!(tx_input_query_response.data.len(), 2);
+
+        // query tx with move call
+        let move_call_query =
+            SuiTransactionResponseQuery::new_with_filter(TransactionFilter::MoveFunction {
+                package: package_id,
+                module: Some("devnet_nft".to_string()),
+                function: None,
+            });
+        let tx_move_call_query_response = indexer_rpc_client
+            .query_transactions(move_call_query, None, None, None)
+            .await?;
+        assert_eq!(tx_move_call_query_response.data.len(), 1);
+        assert_eq!(
+            tx_move_call_query_response.data.first().unwrap().digest,
+            nft_digest
+        );
+
+        Ok(())
+    }
 
     #[tokio::test]
     async fn test_multi_get_transactions_order() -> Result<(), anyhow::Error> {
@@ -355,7 +387,7 @@ pub mod pg_integration_test {
         wait_until_transaction_synced(&store, nft_digest.base58_encode().as_str()).await;
 
         let tx_multi_read_tx_response_1 = indexer_rpc_client
-            .multi_get_transactions_with_options(
+            .multi_get_transactions(
                 vec![tx_response.digest, nft_digest],
                 Some(SuiTransactionResponseOptions::full_content()),
             )
@@ -365,7 +397,7 @@ pub mod pg_integration_test {
         assert_eq!(tx_multi_read_tx_response_1[1].digest, nft_digest);
 
         let tx_multi_read_tx_response_2 = indexer_rpc_client
-            .multi_get_transactions_with_options(
+            .multi_get_transactions(
                 vec![nft_digest, tx_response.digest],
                 Some(SuiTransactionResponseOptions::full_content()),
             )
@@ -389,11 +421,8 @@ pub mod pg_integration_test {
 
         let (_, _, digest_one) = create_devnet_nft(context, package_id).await.unwrap();
         wait_until_transaction_synced(&store, digest_one.base58_encode().as_str()).await;
-        let (_, _, digest_two) = create_devnet_nft(context, package_id).await.unwrap();
+        let (sender, _, digest_two) = create_devnet_nft(context, package_id).await.unwrap();
         wait_until_transaction_synced(&store, digest_two.base58_encode().as_str()).await;
-        let (transferred_object, sender, receiver, digest_three, _, _) =
-            transfer_coin(context).await.unwrap();
-        wait_until_transaction_synced(&store, digest_three.base58_encode().as_str()).await;
 
         // Test various ways of querying events
         let filter_on_sender = EventFilter::Sender(sender);
@@ -440,30 +469,6 @@ pub mod pg_integration_test {
         assert_eq!(digest_one, query_response.data[0].id.tx_digest);
         assert_eq!(digest_two, query_response.data[1].id.tx_digest);
 
-        // Verify that the transfer coin event occurred successfully, without emitting an event
-        let object_correctly_transferred = indexer_rpc_client
-            .get_owned_objects(
-                receiver,
-                Some(SuiObjectResponseQuery::new_with_options(
-                    SuiObjectDataOptions::full_content(),
-                )),
-                None,
-                None,
-                None,
-            )
-            .await?
-            .data
-            .into_iter()
-            .filter_map(|object_resp| {
-                if let Some(data) = object_resp.data {
-                    Some(data.object_id)
-                } else {
-                    None
-                }
-            })
-            .any(|obj| obj == transferred_object);
-        assert!(object_correctly_transferred);
-
         Ok(())
     }
 
@@ -481,7 +486,7 @@ pub mod pg_integration_test {
             let (sender, object_id, digest) = create_devnet_nft(context, package_id).await.unwrap();
             wait_until_transaction_synced(&store, digest.base58_encode().as_str()).await;
             let obj_resp = indexer_rpc_client
-                .get_object_with_options(object_id, None)
+                .get_object(object_id, None)
                 .await
                 .unwrap();
             let data = obj_resp.object()?;
@@ -549,7 +554,7 @@ pub mod pg_integration_test {
             show_storage_rebate: true,
         };
         let resp = indexer_rpc_client
-            .get_object_with_options(source_object_id, Some(show_all_content.clone()))
+            .get_object(source_object_id, Some(show_all_content.clone()))
             .await
             .unwrap();
         let initial_full_obj_data = resp.object()?;
@@ -565,7 +570,7 @@ pub mod pg_integration_test {
         wait_until_transaction_synced(&store, tx_response.digest.base58_encode().as_str()).await;
 
         let response = indexer_rpc_client
-            .get_object_with_options(source_object_id, Some(show_all_content.clone()))
+            .get_object(source_object_id, Some(show_all_content.clone()))
             .await?;
         let post_transfer_full_obj_data = response.object()?;
         let object_required_fields = SuiObjectData {
@@ -580,37 +585,30 @@ pub mod pg_integration_test {
         };
         let show_some_content = SuiObjectDataOptions::new();
         let futures = vec![
-            indexer_rpc_client.get_object_with_options(
-                source_object_id,
-                Some(SuiObjectDataOptions::bcs_lossless()),
-            ),
-            indexer_rpc_client.get_object_with_options(
-                source_object_id,
-                Some(SuiObjectDataOptions::full_content()),
-            ),
             indexer_rpc_client
-                .get_object_with_options(source_object_id, Some(show_some_content.clone())),
-            indexer_rpc_client.get_object_with_options(
+                .get_object(source_object_id, Some(SuiObjectDataOptions::bcs_lossless())),
+            indexer_rpc_client
+                .get_object(source_object_id, Some(SuiObjectDataOptions::full_content())),
+            indexer_rpc_client.get_object(source_object_id, Some(show_some_content.clone())),
+            indexer_rpc_client.get_object(
                 source_object_id,
                 Some(show_some_content.clone().with_content()),
             ),
-            indexer_rpc_client.get_object_with_options(
+            indexer_rpc_client.get_object(
                 source_object_id,
                 Some(show_some_content.clone().with_owner()),
             ),
-            indexer_rpc_client.get_object_with_options(
+            indexer_rpc_client.get_object(
                 source_object_id,
                 Some(show_some_content.clone().with_type()),
             ),
-            indexer_rpc_client.get_object_with_options(
+            indexer_rpc_client.get_object(
                 source_object_id,
                 Some(show_some_content.clone().with_display()),
             ),
-            indexer_rpc_client.get_object_with_options(
-                source_object_id,
-                Some(show_some_content.clone().with_bcs()),
-            ),
-            indexer_rpc_client.get_object_with_options(
+            indexer_rpc_client
+                .get_object(source_object_id, Some(show_some_content.clone().with_bcs())),
+            indexer_rpc_client.get_object(
                 source_object_id,
                 Some(show_some_content.clone().with_previous_transaction()),
             ),
@@ -698,7 +696,7 @@ pub mod pg_integration_test {
         .await?;
         wait_until_transaction_synced(&store, tx_response.digest.base58_encode().as_str()).await;
         let resp = indexer_rpc_client
-            .get_object_with_options(post_transfer_full_obj_data.object_id, None)
+            .get_object(post_transfer_full_obj_data.object_id, None)
             .await
             .unwrap();
 
@@ -726,7 +724,7 @@ pub mod pg_integration_test {
         // Not exists
         let obj_id = ObjectID::from([42; 32]);
         let resp = indexer_rpc_client
-            .get_object_with_options(obj_id, Some(show_all_content.clone()))
+            .get_object(obj_id, Some(show_all_content.clone()))
             .await
             .unwrap();
 
@@ -974,7 +972,7 @@ pub mod pg_integration_test {
             execute_simple_transfer(&mut test_cluster, &indexer_rpc_client).await?;
         wait_until_transaction_synced(&store, tx_response.digest.base58_encode().as_str()).await;
         let full_transaction_response = indexer_rpc_client
-            .get_transaction_with_options(
+            .get_transaction(
                 tx_response.digest,
                 Some(SuiTransactionResponseOptions::full_content()),
             )
@@ -993,9 +991,7 @@ pub mod pg_integration_test {
         ];
         let futures = sui_transaction_response_options
             .into_iter()
-            .map(|option| {
-                indexer_rpc_client.get_transaction_with_options(tx_response.digest, Some(option))
-            })
+            .map(|option| indexer_rpc_client.get_transaction(tx_response.digest, Some(option)))
             .collect::<Vec<_>>();
 
         let received_transaction_results: Vec<SuiTransactionResponse> = join_all(futures)
@@ -1065,7 +1061,7 @@ pub mod pg_integration_test {
         wait_until_transaction_synced(&store, tx_response.digest.base58_encode().as_str()).await;
         // We do this as checkpoint field is only returned in the read api
         let tx_response = indexer_rpc_client
-            .get_transaction_with_options(
+            .get_transaction(
                 tx_response.digest,
                 Some(SuiTransactionResponseOptions::full_content()),
             )
