+++ crates/sui-open-rpc/src/generate_json_rpc_spec.rs
@@ -1,18 +1,17 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use std::fs::File;
-use std::io::Write;
-
+use crate::examples::RpcExampleProvider;
 use clap::ArgEnum;
 use clap::Parser;
 use pretty_assertions::assert_str_eq;
+use std::fs::File;
+use std::io::Write;
 use sui_core::SUI_CORE_VERSION;
-
-use crate::examples::RpcExampleProvider;
 use sui_json_rpc::api::ExtendedApiOpenRpc;
+use sui_json_rpc::api::IndexerApiOpenRpc;
+use sui_json_rpc::api::MoveUtilsOpenRpc;
 use sui_json_rpc::coin_api::CoinReadApi;
-use sui_json_rpc::event_api::EventReadApi;
 use sui_json_rpc::governance_api::GovernanceReadApi;
 use sui_json_rpc::read_api::ReadApi;
 use sui_json_rpc::sui_rpc_doc;
@@ -48,11 +47,12 @@ async fn main() {
     let mut open_rpc = sui_rpc_doc(SUI_CORE_VERSION);
     open_rpc.add_module(ReadApi::rpc_doc_module());
     open_rpc.add_module(CoinReadApi::rpc_doc_module());
-    open_rpc.add_module(EventReadApi::rpc_doc_module());
+    open_rpc.add_module(IndexerApiOpenRpc::module_doc());
     open_rpc.add_module(TransactionExecutionApi::rpc_doc_module());
     open_rpc.add_module(TransactionBuilderApi::rpc_doc_module());
     open_rpc.add_module(GovernanceReadApi::rpc_doc_module());
     open_rpc.add_module(ExtendedApiOpenRpc::module_doc());
+    open_rpc.add_module(MoveUtilsOpenRpc::module_doc());
 
     open_rpc.add_examples(RpcExampleProvider::new().examples());
