+++ narwhal/primary/src/proposer.rs
@@ -477,7 +477,7 @@ impl Proposer {
                     self.metrics
                         .proposal_latency
                         .with_label_values(&[reason])
-                        .observe(((current_timestamp - t) / 1000) as f64);
+                        .observe(Duration::from_millis(current_timestamp - t).as_secs_f64());
                 }
                 self.last_round_timestamp = Some(current_timestamp);
                 debug!("Dag moved to round {}", self.round);
